# Improved the latency of ResultGears by skipping middle frames.