* The new implementation analyzes only the first and last frame.
* It improves the latency more than 25 seconds in my machine (MacBook Air 2014)

Durations of analysis against the same result scene.

Before:
  42.41s user 8.86s system 110% cpu 46.364 total

After:
  14.63s user 1.83s system 111% cpu 14.702 total